properly merge labels and jumps
Labels are filtered out of the bytecode and kept separetely (only
labels that are target of jump instructions). The rest is then
diffed once, considering all labels (and thus jumps) different.
From the first diff, labels that are in the same place are noted
and a second diff is run that uses the label information to
consider equal jumps to equal labels.

When pretty printing, the number of the instruction following the
label is used, preferably using the one from the target file (this
needs improvement; it should be possible to differentiate visually
the old/new jump target)