another step towards proper label merging.
We now filter out all input labels that are not jump targets (they
are useless for us), and support label-aware diffing for both
jump instructions and lookupswitch.

Still remains the main display issue: there is currently no way
to know when looking at the output if a target ins. number
references the first file or the second file. Results are
sometime surprising, but it only affects the display, not whether
two instruction are different or not.